+++ changelog.md
@@ -2,6 +2,7 @@
 
 - **0.1.1**
     - Resolves RAMON download bug
+    - Adds `ramon_type` filter to `remote.neurodata.get_ramon_ids`
 - **0.1.0**
     - Deprecated OCP in favor of ndio.remote.neurodata
 - **0.0.24**